4.23

+++ osa004/bloglist/controllers/blogs.js
@@ -1,64 +1,87 @@
 const blogsRouter = require('express').Router();
 const Blog = require('../models/blog');
+const User = require('../models/user');
 
-blogsRouter.get('/', (request, response) => {
-  Blog.find({}).then(blogs => {
-    response.json(blogs);
-  });
+blogsRouter.get('/', async (request, response) => {
+  const blogs = await Blog
+    .find({}).populate('user', { username: 1, name: 1 });
+  response.json(blogs);
 });
 
-blogsRouter.get('/:id', (request, response, next) => {
-  Blog.findById(request.params.id)
-    .then(blog => {
-      if (blog) {
-        response.json(blog);
-      } else {
-        response.status(404).end();
-      }
-    })
-    .catch(error => next(error));
+blogsRouter.get('/:id', async (request, response) => {
+  const blog = await Blog.findById(request.params.id)
+    .populate('user', { username: 1, name: 1 });
+
+  if (blog)
+    response.json(blog);
+  else
+    response.status(404).end();
 });
 
-blogsRouter.post('/', (request, response, next) => {
+blogsRouter.post('/', async (request, response) => {
   const body = request.body;
+
+  let user = request.user;
+
+  if (!user)
+    user = await User.findOne({});
 
   const blog = new Blog({
     title: body.title,
     author: body.author,
     url: body.url,
-    likes: body.likes
+    likes: body.likes,
+    user: user._id
   });
 
-  blog.save()
-    .then(savedBlog => {
-      response.json(savedBlog);
-    })
-    .catch(error => next(error));
+  const savedBlog = await blog.save();
+  user.blogs = user.blogs.concat(savedBlog._id);
+  await user.save();
+
+  response.status(201).json(savedBlog);
 });
 
-blogsRouter.delete('/:id', (request, response, next) => {
-  Blog.findByIdAndDelete(request.params.id)
-    .then(() => {
-      response.status(204).end();
-    })
-    .catch(error => next(error));
+blogsRouter.delete('/:id', async (request, response) => {
+  const userFromToken = request.user;
+
+  if (!userFromToken)
+    return response.status(401).json({ error: 'user not found' });
+
+  const blog = await Blog.findById(request.params.id);
+  if (!blog)
+    return response.status(404).json({ error: 'blog not found' });
+
+  if (blog.user.toString() !== userFromToken._id.toString())
+    return response.status(403).json({ error: 'only the creator can delete this blog' });
+
+  await Blog.findByIdAndDelete(request.params.id);
+  response.status(204).end();
 });
 
-blogsRouter.put('/:id', (request, response, next) => {
+blogsRouter.put('/:id', async (request, response) => {
   const body = request.body;
 
-  const blog = {
+  const userFromToken = request.user;
+
+  if (!userFromToken)
+    return response.status(401).json({ error: 'user not found' });
+
+  const blog = await Blog.findById(request.params.id);
+  if (!blog)
+    return response.status(404).json({ error: 'blog not found' });
+
+  if (blog.user.toString() !== userFromToken._id.toString())
+    return response.status(403).json({ error: 'only the creator can modify this blog' });
+
+  const blogWithUpdatedData = {
     title: body.title,
     author: body.author,
     url: body.url,
     likes: body.likes
   };
 
-  Blog.findByIdAndUpdate(request.params.id, blog, { new: true })
-    .then(updatedBlog => {
-      response.json(updatedBlog);
-    })
-    .catch(error => next(error));
+  const updatedBlog = await Blog.findByIdAndUpdate(request.params.id, blogWithUpdatedData, { new: true });
+  response.json(updatedBlog);
 });
 
 module.exports = blogsRouter;
